+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.7.0] - 2025-04-16
+### Added
+- Support for dynamic commands with system actions (not the same as key mapping)
+- Create command function integrated with the WebView
+- Error handling to avoid duplicated redirected keys or commands in one profile
+### Modified
+
+
 ## [0.6.3] - 2025-04-16
 ### Added
 - First draft of default profile